Update Grove-Multichannel_Gas_Sensor.md
This commit adds a comment explicitly mentioning that the I2C address for the Grove Multichannel Gas Sensor v2.0 is 0x08. This clarification ensures that users correctly initialize the sensor when using the gas.begin(0x08); function in their code.

The update helps avoid confusion, as different versions of the sensor may have varying default I2C addresses. This  improvement makes the code more readable and easier to use, especially for those unfamiliar with the sensor’s communication settings.